+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### API-Changes
 
 ### Changes
+- allow sender timestamp to be in the future, but not too much
 
 ### Fixes
 - `dc_search_msgs()` returns unaccepted requests #3694